# Tweak vertex lighter to avoid causing excess array initialisations