When looking at some sampling results, I noticed `QuadGatheringTransformer.setVertexFormat()` showed up as requiring a noticable amount of CPU time.

Looking at the code, the main thing this method does is allocate new `int[]` and `float[][][]` arrays to hold quad data.

`setVertexFormat()` is called from `VertexLighterFlat.setParent()`, if the vertex format is not equal to the parent's format.

However, the actual format that is set (via static method `getVertexFormat(...)`) will not necessarily be the same as the parent format, a `NORMAL` element will be appended if missing from the parent format.

This means that for vertex formats without a normal, the current equality check will always fail, leading to the following code running even though the format used has not changed.

This PR alters the check to compare against the format that would be passed to `setVertexFormat()` instead, preventing unneeded code running and the resulting object creation/disposal overhead.