@bs2609 2017-03-10
@RainWarrior, as the authority on this code, can you take a look at this when you have time?