@bs2609 2017-03-10
It also might be better to not create new `Format` and `FormatElement` objects each time as well.