@bs2609 2017-05-01
Here's some screenshots of sampling snapshots I took with VisualVM.

* Before changes: (regular build 2282)
![mod_before](https://cloud.githubusercontent.com/assets/1447117/25579492/569a128e-2e70-11e7-91e5-93a49b8069fc.png)

* After changes: (dev build)
![mod_after](https://cloud.githubusercontent.com/assets/1447117/25579504/6ce5964e-2e70-11e7-909f-b713e6723b96.png)