@RainWarrior 2017-05-01
Would you mind doing the profiling in addition to the sampling? I've found that sampling doesn't always yield accurate results :)