@mezz 2017-05-01
Making my own profiling setup to confirm:
1. Create a new world, go to a jungle, and wait for chunk updates to die down.
2. Start profiling and hit F3+A a few times to re-render everything then stop profiling.
3. Exit game, switch to PR branch and repeat step 2.

Results, looking at `ForgeBlockModelRenderer.renderModelSmooth`:
Before the PR, `VertexLighterFlat.setParent` wasn't taking a ton of time but it did show up. It takes about 1/10 as much time as `ForgeBlockModelRenderer.render`.
After the PR, `setParent` barely shows up, taking about 1/500 as much time as `render`.

Looks good to me, 8-10% is a bigger performance boost than I was expecting. Since the chunk re-rendering is threaded and my computer isn't struggling it's not super noticeable, but definitely not bad.