@bs2609 2017-05-02
After some hassle, I was able to get some profiling data.

Here's the results I got from basically the same test environment as in the sampling setup.

Profiling this code made a significant difference to general performance, so it took a relatively long time to get data for any real amount of 'work'. Also note that these results are dependent on calibration data, so might not be fully correct themselves.

Method invocation count applies to all methods listed apart from where noted.

*  Without changes: (8,057,272 method invocations recorded)
    *  `QuadGatheringTransformer#setVertexFormat(VertexFormat)`: 10,798ms self, 11,182ms total
    *  `VertexLighterFlat#setParent(IVertexConsumer)`: 5,109ms self, 35,724ms total
    *  `ForgeBlockModelRenderer.render(...)`: 4,565ms self, 282,765ms total
    *  `ForgeBlockModelRenderer#renderModelSmooth(...)`: 3,114ms self, 323,490ms total
    *  `VertexLighterFlat.getVertexFormat(IVertexConsumer)`: 1,657ms self, 16,082ms total

*  With changes: (7,547,688 method invocations recorded)
    *  `QuadGatheringTransformer#setVertexFormat(VertexFormat)`: ~0ms self, ~0ms total (very few invocations)
    *  `VertexLighterFlat#setParent(IVertexConsumer)`: 874ms self, 14,019ms total
    *  `ForgeBlockModelRenderer.render(...)`: 3,862ms self, 241,087ms total
    *  `ForgeBlockModelRenderer#renderModelSmooth(...)`: 2,929ms self, 259,633ms total
    *  `VertexLighterFlat.getVertexFormat(IVertexConsumer)`: 1,071ms self, 10,554ms total